collapse two files into one and clean up comments
